Rearrange internals of ramutils.tasks
Moves task decorator to allow importing all tasks at once.

Closes #30.
