+++ ramutils/tasks/misc.py
@@ -1,0 +1,15 @@
+import os
+from ptsa.data.readers import JsonIndexReader
+from ._wrapper import task
+
+
+@task()
+def read_index(mount_point='/'):
+    """Reads the JSON index reader.
+
+    :param str mount_point: Root directory to search for.
+    :returns: JsonIndexReader
+
+    """
+    path = os.path.join(mount_point, 'protocols', 'r1.json')
+    return JsonIndexReader(path)
